fix(x64): widen in-memory gold paths; refined residual warning pass
- Redefine (not cast) where the whole path is in-memory 64-bit: CItem::
  GetGold()/GetShopBuyPrice() -> GoldType (proto fields are GoldType,
  consumers take GoldType); core_dump_unix line param WORD -> int.
- config.cpp guild_create_fee: keep int (a fee never nears 1e17) but
  drop the nonsensical GOLD_MAX ceiling that didn't fit int -- clamp to
  >=1 instead (this was the real C4305/C4309 cause, not the type).
- Refined parallel pass: time_t->32-bit timestamps cast explicitly
  (protocol/DB store 32-bit), liblua bounded index/size casts, generic
  bounded size_t casts.

Server x64: 0 compile errors; warnings ~326 -> ~130. Remaining are
boundary casts into 32-bit DB/packet/item-count contracts, a few
std::map type mismatches, the qc tool .cc files, and two flagged items
needing review (switchbot BYTE packet field; a qc '==' vs '=' bug).

+++ Lead-Server-Source/db/src/ClientManager.cpp
@@ -699,9 +699,9 @@ void CClientManager::RESULT_SAFEBOX_LOAD(CPeer * pkPeer, SQLMsg * msg)
 								if (LIMIT_REAL_TIME == item_table->aLimits[i].bType)
 								{
 									if (0 == item_table->aLimits[i].lValue)
-										pItemAward->dwSocket0 = time(0) + 60 * 60 * 24 * 7;
+										pItemAward->dwSocket0 = static_cast<DWORD>(time(0) + 60 * 60 * 24 * 7);
 									else
-										pItemAward->dwSocket0 = time(0) + item_table->aLimits[i].lValue;
+										pItemAward->dwSocket0 = static_cast<DWORD>(time(0) + item_table->aLimits[i].lValue);
 
 									break;
 								}
