+++ Lead-Server-Source/game/src/guild_manager.cpp
@@ -702,7 +702,7 @@ bool CGuildManager::EndWar(DWORD guild_id1, DWORD guild_id2)
 	if (g2)
 		g2->EndWar(guild_id1);
 
-	m_GuildWarEndTime[k] = get_global_time();
+	m_GuildWarEndTime[k] = static_cast<DWORD>(get_global_time());
 	CHARACTER_MANAGER::instance().for_each_pc(FSendWarList(GUILD_SUBHEADER_GC_GUILD_WAR_END_LIST, guild_id1, guild_id2));
 	m_GuildWar.erase(it);
